component for ActivityBar, handle body scroll on modal
